Move Config class out of gui
It is used from non-gui code too
